Feature module extraction: 7 feature modules per sprint pack
Extract all main screens from app/ into dedicated feature modules:

- feature-home: HomeScreen + HomeViewModel
- feature-map: MapScreen + MapViewModel
- feature-dialogue: DialogueSceneScreen + DialogueSceneViewModel
- feature-camp: CampScreen + CampViewModel
- feature-journal: JournalScreen + JournalViewModel
- feature-party: PartyScreen + PartyViewModel
- feature-settings: SettingsScreen + SettingsViewModel

Each feature module:
- Has its own build.gradle.kts with Compose, Hilt, Navigation
- Depends on core-model, core-ui, core-database, core-data, domain
- Uses package com.chimera.feature.{name}
- Has AndroidManifest.xml

app/ retains only navigation-level screens: duel, splash, onboarding,
saveslot (these are entry points, not tab destinations).

Total module count: 15
  8 core/domain modules + 7 feature modules
