+++ settings.gradle.kts
@@ -22,3 +22,10 @@ include(":core-network")
 include(":core-ai")
 include(":core-data")
 include(":domain")
+include(":feature-home")
+include(":feature-map")
+include(":feature-dialogue")
+include(":feature-camp")
+include(":feature-journal")
+include(":feature-party")
+include(":feature-settings")
